+++ Projects/Domain/UseCase/Sources/DeparturePlace/SetDefaultDeparturePlaceUseCase.swift
@@ -6,5 +6,5 @@
 import Entity
 
 public protocol SetDefaultDeparturePlaceUseCase: Sendable {
-    func execute(id: Int64) async throws -> DeparturePlace
+    func execute(id: Int) async throws -> DeparturePlace
 }
